Yes. I made concrete speed-focused changes in `EInkDisplay.cpp` and verified it still builds.
**What I optimized**
- Batched many tiny SPI data writes into block writes in init/RAM area setup:
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:239`
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:270`
- Increased block-send helper length to `uint32_t` so large transfers stay native:
  - `open-x4-sdk/libs/display/EInkDisplay/include/EInkDisplay.h:94`
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:203`
- Optimized `drawImage` to clip once and use `memcpy` for non-PROGMEM paths:
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:309`
- Optimized `drawImageTransparent` similarly (single clip pass, tighter loop, guarded logging):
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:345`
- Removed redundant RED RAM full-buffer write for non-fast refresh (RED is bypassed there):
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:458`
- Reworked `displayWindow` to stream rows directly to SPI (no `std::vector` alloc/copy):
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:515`
- Batched custom LUT upload (105 bytes in one transfer instead of 105 single-byte sends):
  - `open-x4-sdk/libs/display/EInkDisplay/src/EInkDisplay.cpp:627`

**Validation**
- Build passed: `pio run` (from `/Users/ben/crosspoint-reader`), success.

+++ libs/display/EInkDisplay/include/EInkDisplay.h
@@ -91,7 +91,7 @@ class EInkDisplay {
   void resetDisplay();
   void sendCommand(uint8_t command);
   void sendData(uint8_t data);
-  void sendData(const uint8_t* data, uint16_t length);
+  void sendData(const uint8_t* data, uint32_t length);
   void waitWhileBusy(const char* comment = nullptr);
   void initDisplayController();
